requirements.txt
supabase==2.8.0
python-dotenv==1.0.0
Pillow==10.4.0
requests==2.31.0
notion-client==2.5.0
opencv-python==4.12.0.88
moviepy==1.0.3
numpy==2.2.6
scikit-learn==1.5.2
transformers==4.46.3
torch==2.6.0
SpeechRecognition==3.10.0
pydub==0.25.1
